cp-291: + handleUnauthorized middleware
