+++ frontend/src/libs/packages/store/libs/enums/enums.ts
@@ -1,0 +1,1 @@
+export { ExceptionMessage } from 'shared/build/index.js';
